Fix colors on FreeBSD (and MacOS?) again
Broken in 1548f9276f48b9d7a378b2e491609fc947f93e6c because I'm an idiot.

Probably also fixes #210, but I don't have a Mac to test.

+++ src/ncurses_refs.c
@@ -27,4 +27,4 @@ chtype ncdu_acs_hline()    { return ACS_VLINE   ; }
 chtype ncdu_acs_vline()    { return ACS_HLINE   ; }
 
 /* https://github.com/ziglang/zig/issues/8947 */
-void ncdu_init_pair(short a, short b, short c) { init_pair(a,b,c); }
+void ncdu_init_pair(int a, int b, int c) { init_pair(a,b,c); }
